[svn-r24388] switch to using Mercury 64 bit CRC algorithm instead of internal checksum algorithm.

+++ src/H5VLiod.h
@@ -30,14 +30,13 @@
 #include "mercury.h"
 #include "mercury_handler.h"
 #include "mercury_macros.h"
-#include "mercury_proc.h"
+//#include "mercury_proc.h"
 #include "mercury_proc_string.h"
-#include "mercury_error.h"
-#include "mercury_bulk.h"
-#include "mercury_config.h"
+//#include "mercury_error.h"
+//#include "mercury_bulk.h"
+//#include "mercury_config.h"
 #include "na_mpi.h"
-#include "na.h"
-
+//#include "na.h"
 
 #define H5VL_IOD	(H5VL_iod_init())
 #else
@@ -62,10 +61,10 @@ H5_DLL herr_t EFF_start_server(MPI_Comm comm, MPI_Info info);
 H5_DLL herr_t EFF_init(MPI_Comm comm, MPI_Info info);
 H5_DLL herr_t EFF_finalize(void);
 
-H5_DLL herr_t H5Pset_dxpl_checksum(hid_t dxpl_id, uint32_t value);
-H5_DLL herr_t H5Pget_dxpl_checksum(hid_t dxpl_id, uint32_t *value);
-H5_DLL herr_t H5Pset_dxpl_checksum_ptr(hid_t dxpl_id, uint32_t *value);
-H5_DLL herr_t H5Pget_dxpl_checksum_ptr(hid_t dxpl_id, uint32_t **value);
+H5_DLL herr_t H5Pset_dxpl_checksum(hid_t dxpl_id, uint64_t value);
+H5_DLL herr_t H5Pget_dxpl_checksum(hid_t dxpl_id, uint64_t *value);
+H5_DLL herr_t H5Pset_dxpl_checksum_ptr(hid_t dxpl_id, uint64_t *value);
+H5_DLL herr_t H5Pget_dxpl_checksum_ptr(hid_t dxpl_id, uint64_t **value);
 H5_DLL herr_t H5Pset_metadata_integrity_scope(hid_t fapl_id, uint32_t scope);
 H5_DLL herr_t H5Pget_metadata_integrity_scope(hid_t fapl_id, uint32_t *scope);
 H5_DLL herr_t H5Pset_rawdata_integrity_scope(hid_t dxpl_id, uint32_t scope);
